+++ HISTORY.rst
@@ -7,6 +7,8 @@ Pending Release
 
 .. Insert new release notes below this line
 
+* Add rule ``I202`` to ban relative imports, when activated with the new
+  ``ban-relative-imports`` configuration option.
 * Update Python support to 3.5-3.7, as 3.4 has reached its end of life.
 
 2.0.0 (2019-02-02)